Sort values numerically before combining with headers
Values are sometimes not returned from Spout in the correct column order.

This fixes this issue by sorting the columns before combining them with headers.

The issue occurs when columns are omitted from the underlying XLSX data. As in this example:
```xml
<row r="3">
    <c r="A3" s="5">
        <v>1.1112222E7</v>
    </c>
    <c r="C3" s="4" t="s">
        <v>4</v>
    </c>
</row>
```
As we see, B3 is not present in this file which makes Spout return columns 1 and 3 as they are in that order in the file. Spout then appends column 2 to the end with a null value. That way, all columns are present in the returned array, but not in the column order, but as 0,2,1 indexes.

Fixes #14